IotKernel/requirements.txt
fastapi
uvicorn
pytest
requests
paho-mqtt
redis